Moving action creator into separate file, amending jsdoc
